Feature/333 v210 packed yuv format

+++ YUViewLib/src/parser/common/SubByteReaderLogging.cpp
@@ -187,7 +187,7 @@ void SubByteReaderLogging::removeLogSubLevel()
 }
 
 uint64_t
-SubByteReaderLogging::readBits(const std::string &symbolName, int numBits, const Options &options)
+SubByteReaderLogging::readBits(const std::string &symbolName, size_t numBits, const Options &options)
 {
   try
   {
